issue #115 rename : model -> model/use_case
