+++ client/app/src/main/java/com/example/witt/domain/use_case/validate/ValidatePhoneNum.kt
@@ -1,6 +1,6 @@
 package com.example.witt.domain.use_case.validate
 
-import com.example.witt.domain.model.validate.ValidationResult
+import com.example.witt.domain.model.use_case.validate.ValidationResult
 
 class ValidatePhoneNum {
